source-code/requirements.txt
# JARVIS - Voice-to-Text System Requirements
# Production-ready dependencies with version pinning

# Core ML/ASR
# NeMo 2.1.0 supports Canary-1B Flash (requires >= r1.23.0)
nemo_toolkit[asr]==2.1.0
torch==2.5.1
torchaudio>=2.6.0

# Audio Processing
sounddevice==0.5.1
scipy==1.14.1
librosa==0.10.2
numpy==1.26.4
wave; python_version >= "3.0"

# Keyboard & Typing
pynput==1.7.7
evdev==1.7.1  # For Wayland-compatible keyboard automation
pyperclip==1.9.0  # For clipboard-based text insertion

# Configuration & Environment
python-dotenv==1.0.1
pyyaml==6.0.2

# LLM Integration
requests==2.32.3
claude-agent-sdk==0.1.4  # Claude Code programmatic integration

# Terminal UI
rich==13.0.0
textual>=0.89.0

# Logging & Monitoring (optional)
prometheus-client==0.21.0

# Development Dependencies (optional)
# pytest==8.3.4
# pytest-cov==6.0.0
# black==24.10.0
# ruff==0.7.4
